add: implement delete blog operation specifically for users and set up user extractor middleware

+++ part4/bloglist/controllers/blogs.js
@@ -1,7 +1,6 @@
-const jwt = require("jsonwebtoken")
+const middleware = require("../utils/middleware")
 const blogsRouter = require("express").Router()
 const Blog = require("../models/blog")
-const User = require("../models/user")
 
 // routes
 blogsRouter.get("/", async (request, response) => {
@@ -9,17 +8,9 @@ blogsRouter.get("/", async (request, response) => {
   response.json(blogs)
 })
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", middleware.userExtractor, async (request, response) => {
   const body = request.body
-
-  // verify token and decodes token to object
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: "token invalid" })
-  }
-
-  // find user based on id
-  const user = await User.findById(decodedToken.id)
+  const user = request.user
 
   // return error if title/url missing from request
   if (!body.title || !body.url) {
@@ -53,10 +44,28 @@ blogsRouter.get("/:id", async (request, response) => {
   }
 })
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
-})
+blogsRouter.delete(
+  "/:id",
+  middleware.userExtractor,
+  async (request, response) => {
+    const user = request.user
+
+    // find existing blog based on ID of DELETE request
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+      return response.status(404).json({ error: "blog not found" })
+    }
+
+    // verify user based on token
+    if (blog.user.toString() !== user.id) {
+      return response.status(401).json({ error: "unauthorized user" })
+    }
+
+    // delete blog
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+)
 
 blogsRouter.put("/:id", async (request, response) => {
   const body = request.body
